Refactor csrf package
